refactor(cluster): extract weight update to REQUIRES_NEW service and harden edge cases (#133)
- Extract ClusterInterestWeightService with REQUIRES_NEW propagation
- Use COALESCE(u.weight, 0) for null-safe atomic weight update
- Add resolveSummaryFallback() to handle blank article summaries
- Bind FeedbackType enum directly in request DTO (no valueOf)
- Fix FeedbackRequest JavaDoc (Bean Validation, not Jackson)
- Update ClusterInteractionServiceTest for new service structure
